Split CRI as a separate engine
Note: this won't work without further changes as now Docker will take
over all matching containers, even if there's no metadata for them.
